bug fix in app raffle.
Ajustado erros de lógica na nova implementação de busca de prêmios do sorteio( globo da sorte com prêmios diferentes).

+++ AppSorteioRXPremiado/Assets/_Core/Scripts/Global/ScriptableObjects/GlobeScriptable.cs
@@ -12,7 +12,7 @@ public class GlobeScriptable : ScriptableObject
     public float value;
     [Space]
     public int Winners;
-    public int prizeValue;
+    public float prizeValue;
     public int ballRaffledCount;
     public int possiblesWinnersCount;
     [Space]
